src/main/java/com/zbfan/spring_order/service/impl/ProductCategoryServiceImpl.java
package com.zbfan.spring_order.service.impl;

import com.zbfan.spring_order.dataobject.ProductCategory;
import com.zbfan.spring_order.dataobject.repository.ProductCategoryRepository;
import com.zbfan.spring_order.service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Map;

@Service
public class ProductCategoryServiceImpl implements ProductCategoryService {

    @Autowired
    private ProductCategoryRepository repository;

    @Override
    public ProductCategory findOne(Integer id) {
        return repository.getById(id);
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> catList) {
        return repository.findByCategoryTypeIn(catList);
    }

    @Override
    public ProductCategory save(ProductCategory category) {
        return repository.save(category);
    }
}
